enhance:moving opsramp config from internal/kube to internal/redis
