+++ processor/k8sattributesprocessor/config.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/k8sconfig"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/k8sattributesprocessor/internal/kube"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/k8sattributesprocessor/internal/redis"
 )
 
 // Config defines configuration for k8s attributes processor.
@@ -42,7 +43,7 @@ type Config struct {
 	//Opsramp Metadata Addons Section
 	MetadataAddOn []AddOnMetadata `mapstructure:"metadata_addon"`
 
-	RedisConfig OpsrampRedisConfig `mapstructure:"redis_config"`
+	RedisConfig redis.OpsrampRedisConfig `mapstructure:"redis_config"`
 }
 
 func (cfg *Config) Validate() error {
@@ -336,13 +337,4 @@ type AddOnMetadata struct {
 	Key string `mapstructure:"key"`
 
 	Value string `mapstructure:"value"`
-}
-
-type OpsrampRedisConfig struct {
-	RedisHost   string `mapstructure:"redisHost"`
-	RedisPort   string `mapstructure:"redisPort"`
-	RedisPass   string `mapstructure:"redisPass"`
-	ClusterName string `mapstructure:"clusterName"`
-	ClusterUid  string `mapstructure:"clusterUid"`
-	NodeName    string `mapstructure:"nodeName"`
 }
