+++ processor/k8sattributesprocessor/options.go
@@ -14,6 +14,7 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/k8sconfig"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/k8sattributesprocessor/internal/kube"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/k8sattributesprocessor/internal/metadata"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/k8sattributesprocessor/internal/redis"
 )
 
 const (
@@ -390,11 +391,11 @@ func withAddOnFields(filters ...AddOnMetadata) option {
 	}
 }
 
-func withRedisConfigFields(filters OpsrampRedisConfig) option {
+func withRedisConfigFields(filters redis.OpsrampRedisConfig) option {
 	fmt.Println("The value of filters received  : ", filters)
 
 	return func(p *kubernetesprocessor) error {
-		p.redisConfig = kube.OpsrampRedisConfig(filters)
+		p.redisConfig = filters
 		return nil
 	}
